Refactor move modal component
