+++ src/components/admin/order/realtime/ServedOrderCard.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
 import { expandButtonStyle } from '@styles/buttonStyles';
 import { colFlex, rowFlex } from '@styles/flexStyles';
-import OrderDetailModal from '@components/admin/order/realtime/modal/OrderDetailModal';
+import OrderDetailModal from '@components/admin/order/realtime/modal/order-detail/OrderDetailModal';
 import { areOrdersEquivalent } from '@utils/MemoCompareFunction';
 import useModal from '@hooks/useModal';
 import { RiArrowRightSLine } from '@remixicon/react';
